requirements.txt
torch>=2.4.0
torchaudio>=2.4.0
transformers>=4.52.1
accelerate>=0.30.0
datasets>=2.19.0,<4.0
peft>=0.18.0
soundfile
librosa
pyyaml
openai-whisper
jiwer
wandb
tensorboard
certifi
